Moved About tab to separate package
