+++ profile/src/main/kotlin/com/imashnake/animite/profile/ProfileScreen.kt
@@ -67,6 +67,7 @@ import com.imashnake.animite.core.ui.NestedScrollableContent
 import com.imashnake.animite.core.ui.StatsRow
 import com.imashnake.animite.core.ui.layouts.BannerLayout
 import com.imashnake.animite.profile.dev.internal.ANILIST_AUTH_DEEPLINK
+import com.imashnake.animite.profile.tabs.AboutTab
 import com.ramcosta.composedestinations.annotation.DeepLink
 import com.ramcosta.composedestinations.annotation.Destination
 import com.ramcosta.composedestinations.annotation.RootNavGraph
@@ -268,148 +269,4 @@ enum class ProfileTabs(@StringRes val titleRes: Int) {
     MANGA(R.string.manga),
     FAVOURITES(R.string.favourites),
     STATISTICS(R.string.statistics)
-}
-
-@Composable
-private fun AboutTab(
-    user: User,
-    mediaCollection: User.MediaCollection?,
-    modifier: Modifier = Modifier
-) {
-    val scrollState = rememberScrollState()
-
-    val statsLabelToValue = listOf(
-        stringResource(R.string.total_anime) to user.count?.toString(),
-        stringResource(R.string.days_watched) to user.daysWatched?.let { "%.1f".format(it) },
-        stringResource(R.string.mean_score) to user.meanScore?.let { "%.1f".format(it) }
-    ).filter { it.second != null }
-
-    // TODO: Cleanup this layout.
-    Column(modifier.verticalScroll(scrollState)) {
-        Column(
-            modifier = Modifier.padding(LocalPaddings.current.large),
-            verticalArrangement = Arrangement.spacedBy(LocalPaddings.current.medium)
-        ) {
-            StatsRow(
-                stats = statsLabelToValue,
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(horizontal = LocalPaddings.current.large)
-                    .landscapeCutoutPadding()
-            ) {
-                Text(
-                    text = it.first,
-                    color = MaterialTheme.colorScheme.onBackground,
-                    style = MaterialTheme.typography.labelSmall,
-                    textAlign = TextAlign.Center
-                )
-
-                it.second?.let { value ->
-                    Text(
-                        text = value,
-                        color = MaterialTheme.colorScheme.onBackground,
-                        style = MaterialTheme.typography.displaySmall
-                    )
-                }
-            }
-
-            if (user.genres.isNotEmpty()) Genres(user.genres)
-
-        }
-        // TODO: Why is this not smart-casting?
-        if (!mediaCollection?.namedLists.isNullOrEmpty()) {
-            UserMediaList(mediaCollection!!.namedLists)
-        }
-    }
-}
-
-@Composable
-private fun Genres(
-    genres: List<User.Genre>,
-    modifier: Modifier = Modifier
-) {
-    val barWidthAnimation = remember { Animatable(0f) }
-    val barAlphaAnimation = remember { Animatable(0f) }
-    val barColor = MaterialTheme.colorScheme.primary
-    LaunchedEffect("barAnimation") {
-        launch {
-            barWidthAnimation.animateTo(
-                targetValue = 1f,
-                animationSpec = tween(durationMillis = 1000, delayMillis = 250)
-            )
-        }
-        launch {
-            barAlphaAnimation.animateTo(
-                targetValue = 1f,
-                animationSpec = tween(durationMillis = 1000, delayMillis = 250)
-            )
-        }
-    }
-
-    Text(
-        text = stringResource(R.string.genres),
-        style = MaterialTheme.typography.titleMedium
-    )
-    Row(modifier.height(IntrinsicSize.Max)) {
-        Column(horizontalAlignment = Alignment.End) {
-            genres.forEach {
-                Text(
-                    text = it.genre,
-                    style = m3TextStyles().textStyle.copy(
-                        color = MaterialTheme.colorScheme.onBackground.copy(alpha = 0.74f)
-                    )
-                )
-            }
-        }
-        Column(
-            modifier = Modifier
-                .fillMaxHeight()
-                .padding(
-                    start = LocalPaddings.current.small,
-                    end = LocalPaddings.current.large
-                )
-                .widthIn(max = 250.dp),
-            verticalArrangement = Arrangement.spacedBy(LocalPaddings.current.tiny)
-        ) {
-            val highestCount = genres.maxOf { it.mediaCount }
-            genres.forEach {
-                val weight = it.mediaCount/highestCount.toFloat()
-                Box(
-                    modifier = Modifier
-                        .fillMaxWidth(fraction = weight)
-                        .weight(1f)
-                        .graphicsLayer { alpha = weight * barAlphaAnimation.value }
-                        .drawBehind {
-                            drawRoundRect(
-                                color = barColor,
-                                size = Size(
-                                    width = size.width * barWidthAnimation.value,
-                                    height = size.height
-                                ),
-                                cornerRadius = CornerRadius(size.height)
-                            )
-                        },
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun UserMediaList(
-    lists: List<User.MediaCollection.NamedList>,
-    modifier: Modifier = Modifier
-) {
-    Column(verticalArrangement = Arrangement.spacedBy(LocalPaddings.current.large)) {
-        lists.fastForEach {
-            MediaSmallRow(it.name, it.list, modifier) { media ->
-                MediaSmall(
-                    image = media.coverImage,
-                    label = media.title,
-                    onClick = {},
-                    modifier = Modifier.width(dimensionResource(coreR.dimen.media_card_width))
-                )
-            }
-        }
-    }
 }
